+++ src/data/index.js
@@ -33,5 +33,5 @@ export {default as motivations} from './motivations.json';
 export {default as qualities} from './qualities.json';
 export {default as skills} from './skills.json';
 export {default as settings} from './settings.json';
-export {dataTypes, customDataTypes, vehicleDataTypes, chars, diceOrder, diceNames, modifiableAttributes, newData} from './lists';
+export {dataTypes, customDataTypes, vehicleDataTypes, chars, diceOrder, diceNames, modifiableAttributes} from './lists';
 export {asyncForEach} from './basicFunctions';